% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{collect.h2oplyr_step}
\alias{collect.h2oplyr_step}
\alias{compute.h2oplyr_step}
\alias{as.data.table.h2oplyr_step}
\alias{as.data.frame.h2oplyr_step}
\alias{as_tibble.h2oplyr_step}
\title{Force computation of a lazy data.table}
\usage{
\method{collect}{h2oplyr_step}(x, ...)

\method{compute}{h2oplyr_step}(x, name = unique_name(), ...)

\method{as.data.table}{h2oplyr_step}(x, keep.rownames = FALSE, ...)

\method{as.data.frame}{h2oplyr_step}(x, ...)

\method{as_tibble}{h2oplyr_step}(x, ...)
}
\arguments{
\item{x}{A \link{lazy_dt}}

\item{...}{Arguments used by other methods.}

\item{keep.rownames}{Ignored as dplyr never preseres rownames.}
}
\description{
\itemize{
\item \code{collect()} returns a tibble, grouped if needed
\item \code{compute()} returns a new \link{lazy_dt}
\item \code{as.data.table()} returns a data.table
\item \code{as.data.frame()} returns a data frame
\item \code{as_tibble()} returns a tibble
}
}
