% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtplyr-package.R
\docType{package}
\name{h2oplyr-package}
\alias{h2oplyr}
\alias{h2oplyr-package}
\title{h2oplyr: H2O Back-End for 'dplyr'}
\description{
Provides an H2O backend for 'dplyr'. The goal of 'h2oplyr' 
    is to allow you to write 'dplyr' code that is automatically translated to 
    the equivalent H2O code. As a side-effect, certain basic H2O functions have been 
    modified to accept a subset of data.table type commands.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/caewok/h2oplyr}
  \item Report bugs at \url{https://gitlab.com/caewok/h2oplyr/issues}
}

}
\author{
\strong{Maintainer}: Michael Enion \email{caewok_2000@yahoo.com}

}
