% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o.R
\name{setnames}
\alias{setnames}
\title{Set column names of H2OFrame}
\usage{
setnames(x, ...)
}
\arguments{
\item{x}{H2OFrame}

\item{old}{When new is provided, character names or numeric positions of column names to change. When new is not provided, a function or the new column names. If a function, it will be called with the current column names and is supposed to return the new column names. The new column names must be the same length as the number of columns. See examples.}

\item{new}{Optional. It can be a function or the new column names. If a function, it will be called with old and expected to return the new column names. The new column names must be the same length as columns provided to old argument.}

\item{skip_absent}{Skip items in old that are missing (i.e. absent) in 'names(x)'. Default FALSE halts with error if any are missing.}
}
\value{
The renamed H2OFrame.
}
\description{
Mimics functionality of \code{\link[data.table]{setnames}} from the data.table package.
Changes the column names of the provided H2OFrame.
}
