% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o.R
\name{[.H2OFrame}
\alias{[.H2OFrame}
\title{Extract parts of the H2OFrame object.}
\usage{
\method{[}{H2OFrame}(data, row, col, drop = TRUE, by, keyby)
}
\description{
Operator to extract parts of the H2OFrame object. This implementation modifies the
\code{\link[h2o]{`[.H2OFrame`}} version to more closely approximate data.table usage.
}
\details{
This implementation modifies the
\code{\link[h2o]{`[.H2OFrame`}} version to more closely approximate data.table usage in the following ways.
}
\section{For column specifications}{

\itemize{
\item Permits use of a list() or .() as the column entry.
\item When using list() or .(), references can be directly to column names. For example, \code{mtcars.hex[, .(cyl, mpg)]}.
\item When using list() or .(), common h2o aggregation functions can be specified without using the "h2o." prefix.
For example, \code{mtcars.hex[, .(mean(cyl))]} instead of \code{mtcars.hex[, .(h2o.mean(cyl))]}.
\item To use a variable, specify it outside the list or use a .env pronoun.
For example, \code{mtcars.hex[, "cyl"]} or \code{mtcars.hex[, c("cyl", "hp")]} or
\code{mtcars.hex[, .(.env$var)]} or \code{mtcars.hex[, var]}.
\item Permits use of :=. See \code{\link[data.table]{.SD}}.
\item Permits renaming columns by passing a named list. For example, \code{mtcars.hex[, .(cyl2 = cyl * 2)]}.
}
}

\section{For row specifications}{

\itemize{
\item Uses references to columns directly. For example, \code{mtcars.hex[cyl == 4, ]}.
\item To use a variable, use a .env pronoun or !! or !!!.
}
}

\section{For group specifications}{

\itemize{
\item Permits use of .SD. See \code{\link[data.table]{.SD}}.
\item Permits use of .N. See \code{\link[data.table]{.N}}.
}
}

