% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o.R
\name{[<-.H2OFrame}
\alias{[<-.H2OFrame}
\title{Tidyeval h2o selection}
\usage{
\method{[}{H2OFrame}(data, row, col, ...) <- value
}
\arguments{
\item{data}{H2OFrame}

\item{row}{Row or an expression that evaluates to a row mask.}

\item{col}{One or more column names, joined using either list(), c(), or datatable's '.()' pronoun.}
}
\description{
Extract or replace H2OFrame elements.
}
\details{
Operators to extract or replace parts of H2OFrame objects.
Uses tidyverse non-standard evaluation; replaces underlying H2O operators.
Refer to column names without strings (see examples).
}
\examples{
\dontrun{
h2o.init()
mtcars.hex <- h2o::as.h2o()
mtcars.hex[, .(mpg, cyl)]
mtcars.hex[, list(x = mpg, y = cyl)]
mtcars.hex[1:5,]
mtcars.hex[cyl == 4 & mpg > 20, c(mpg, cyl, wt)]
}
}
